 \name{residuals.va}
\alias{residuals.va}
\title{
Residuals for Variational Approximations
}
\description{
residuals values based on a variational approximation object
}

\usage{
\method{residuals}{va}(object, ...)
}
\arguments{
\item{object}{
 object of va-type}
\item{...}{Additional arguments}
}

\value{
The residuals are returned as a vector}

\author{
Elisabeth Waldmann \cr
\email{ewaldma@uni-goettingen.de}
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
 